# Backend/openapi-v1.yaml
openapi: 3.0.3
info:
  title: GahamBani API
  version: 1.0.0
  description: Documentation de l'API GahamBani. Le Domaine 1 couvre l'authentification
    et la gestion du compte utilisateur.
paths:
...
  /api/v1/auth/change-password/:
    post:
      operationId: auth_change_password
      description: Change the authenticated user's password after validating the old
        one.
      summary: Changer son mot de passe
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/login/:
    post:
      operationId: auth_login
      description: Authentifie un utilisateur par e-mail et retourne une paire de
        jetons JWT.
      summary: Se connecter
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/logout/:
    post:
      operationId: auth_logout
      description: Place le jeton de rafraîchissement fourni sur la liste noire.
      summary: Se déconnecter
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/me/:
    get:
      operationId: auth_me_retrieve
      description: Retrieve or update the authenticated user's permitted profile fields.
      summary: Consulter son compte
      tags:
      - Domaine 1 - Authentification
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
        '401':
          description: Authentification JWT requise ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
    patch:
      operationId: auth_me_update
      description: Modifie uniquement le prénom, le nom et le numéro de téléphone.
      summary: Modifier son compte
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedMeRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedMeRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedMeRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '401':
          description: Authentification JWT requise ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/password-reset/:
    post:
      operationId: auth_password_reset
      description: Envoie un lien de réinitialisation par e-mail si un compte actif
        existe. La réponse reste identique quel que soit le compte demandé.
      summary: Demander une réinitialisation du mot de passe
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordResetResponse'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/password-reset-confirm/:
    post:
      operationId: auth_password_reset_confirm
      description: Set a new password using valid Django reset credentials.
      summary: Confirmer la réinitialisation du mot de passe
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/refresh/:
    post:
      operationId: auth_token_refresh
      description: Retourne un nouveau jeton d’accès à partir d’un jeton de rafraîchissement.
      summary: Rafraîchir un jeton JWT
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedTokenRefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VersionedTokenRefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VersionedTokenRefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
          description: ''
        '401':
          description: Authentification JWT requise ou invalide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
  /api/v1/auth/register/:
    post:
      operationId: auth_register
      description: Crée atomiquement un compte PATIENT actif et non vérifié ainsi
        qu’un profil patient incomplet. Aucun jeton JWT n’est émis.
      summary: Créer un compte patient et son profil
      tags:
      - Domaine 1 - Authentification
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterV1Response'
          description: ''
        '400':
          description: Requête invalide ou erreur de validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '409':
          description: Cette adresse e-mail est déjà utilisée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '429':
          description: Limite de requêtes dépassée. Le délai est indiqué par DRF lorsque
            disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
        '500':
          description: Le compte et le profil ont été annulés car le profil n’a pas
            pu être créé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1Error'
...
components:
  schemas:
...
    ChangePasswordRequest:
      type: object
      description: Validate and change the authenticated user's password.
      properties:
        old_password:
          type: string
          writeOnly: true
          minLength: 1
        new_password:
          type: string
          writeOnly: true
          minLength: 1
        new_password_confirm:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - new_password
      - new_password_confirm
      - old_password
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - password
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
        access:
          type: string
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/LoginUser'
          readOnly: true
      required:
      - access
      - message
      - refresh
      - user
    LoginUser:
      type: object
      description: Expose the public user fields returned after a successful login.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
      required:
      - email
      - first_name
      - id
      - last_name
      - role
    LogoutRequest:
      type: object
      properties:
        refresh:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - refresh
    Me:
      type: object
      description: Serialize the authenticated user and limit editable profile fields.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        phone:
          type: string
          title: Phone number
          maxLength: 30
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
        is_verified:
          type: boolean
          readOnly: true
          title: Verified
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - email
      - id
      - is_verified
      - role
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      required:
      - message
...
    PasswordResetConfirmRequest:
      type: object
      description: Validate a Django reset token and set the user's new password.
      properties:
        uid:
          type: string
          writeOnly: true
          minLength: 1
        token:
          type: string
          writeOnly: true
          minLength: 1
        new_password:
          type: string
          writeOnly: true
          minLength: 1
        new_password_confirm:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - new_password
      - new_password_confirm
      - token
      - uid
    PasswordResetRequest:
      type: object
      description: Send reset credentials by email without revealing account existence.
      properties:
        email:
          type: string
          format: email
          writeOnly: true
          minLength: 1
      required:
      - email
    PasswordResetResponse:
      type: object
      properties:
        detail:
          type: string
          readOnly: true
      required:
      - detail
...
    PatchedMeRequest:
      type: object
      description: Serialize the authenticated user and limit editable profile fields.
      properties:
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        phone:
          type: string
          title: Phone number
          maxLength: 30
...
    RegisterRequest:
      type: object
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 150
        last_name:
          type: string
          minLength: 1
          maxLength: 150
        email:
          type: string
          format: email
          minLength: 1
          title: Email address
          maxLength: 254
        phone:
          type: string
          minLength: 1
          title: Phone number
          maxLength: 30
        password:
          type: string
          writeOnly: true
          minLength: 1
        password_confirm:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - email
      - first_name
      - last_name
      - password
      - password_confirm
      - phone
    RegisterV1Response:
      type: object
      properties:
        detail:
          type: string
          readOnly: true
        user:
          allOf:
          - $ref: '#/components/schemas/RegistrationUser'
          readOnly: true
        profile:
          allOf:
          - $ref: '#/components/schemas/RegistrationProfile'
          readOnly: true
        onboarding:
          allOf:
          - $ref: '#/components/schemas/RegistrationOnboarding'
          readOnly: true
      required:
      - detail
      - onboarding
      - profile
      - user
    RegistrationOnboarding:
      type: object
      properties:
        is_complete:
          type: boolean
          readOnly: true
        completion_percentage:
          type: integer
          readOnly: true
        missing_fields:
          type: array
          items:
            type: string
          readOnly: true
      required:
      - completion_percentage
      - is_complete
      - missing_fields
    RegistrationProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        profile_type:
          type: string
          readOnly: true
      required:
      - id
      - profile_type
    RegistrationUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          readOnly: true
      required:
      - email
      - id
      - role
    RoleEnum:
      enum:
      - PATIENT
      - DOCTOR
      - ADMIN
      type: string
      description: |-
        * `PATIENT` - Patient
        * `DOCTOR` - Doctor
        * `ADMIN` - Admin
...
    TokenRefreshResponse:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
          readOnly: true
      required:
      - access
      - refresh
...
    VersionedTokenRefreshRequest:
      type: object
      description: Reject refresh tokens issued for an older user token version.
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
    V1Error:
      type: object
      required:
      - code
      - detail
      properties:
        code:
          type: string
          example: validation_error
        detail:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        retry_after:
          type: integer
          minimum: 0
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT